+++ header.h
@@ -1,7 +1,8 @@
+
 #ifndef HEADER
 #define HEADER
 #include <string>
-int change(float a, char op, float b);
+float change(float a, char op, float b);
 float bind(std::string a, char opa, std::string b, char opb);
 char check(float f);
 #endif
